+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/newSuits/queries/runjsHappyPath.skip.js
@@ -16,13 +16,13 @@ import {
   query,
   renameQueryFromEditor,
   selectQueryFromLandingPage
-} from "Support/utils/queries";
+} from "Support/utils/appBuilder/querymanager/queries";
 
 import { deleteDownloadsFolder } from "Support/utils/common";
 import {
   resizeQueryPanel,
   verifypreview
-} from "Support/utils/dataSource";
+} from "Support/utils/marketplace/datasources/dataSource";
 import { openNode, verifyNodeData } from "Support/utils/inspector";
 
 describe("RunJS", () => {
