+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/newSuits/table/tableInteractions.cy.js
@@ -19,7 +19,7 @@ import {
 } from "Support/utils/table";
 import { openEditorSidebar, openAccordion } from "Support/utils/commonWidget";
 import { addMultiEventsWithAlert } from "Support/utils/events";
-import { resizeQueryPanel } from "Support/utils/dataSource";
+import { resizeQueryPanel } from "Support/utils/marketplace/datasources/dataSource";
 
 // ---------------------------------------------------------------------------
 // Chunk 3 — Table selection / pagination / sort / filter / search (NewTable).
