+++ cypress-tests/cypress/e2e/happyPath/platform/ceTestcases/dataSources/dataSourcePermissions.cy.skip.js
@@ -3,13 +3,13 @@ import { commonSelectors, commonWidgetSelector } from "Selectors/common";
 import {
   selectAndAddDataSource,
   fillConnectionForm,
-} from "Support/utils/postgreSql";
+} from "Support/utils/marketplace/datasources/postgreSql";
 import { commonText } from "Texts/common";
 import {
   deleteDatasource,
   addQuery,
   verifyValueOnInspector,
-} from "Support/utils/dataSource";
+} from "Support/utils/marketplace/datasources/dataSource";
 import { dataSourceSelector } from "Selectors/dataSource";
 import { dataSourceText } from "Texts/dataSource";
 import {
@@ -19,7 +19,7 @@ import {
 } from "Support/utils/common";
 import { inviteUserBasedOnRole } from "Support/utils/manageGroups";
 import { resolveHost } from "Support/utils/apps";
-import { addSuccessNotification } from "Support/utils/queries";
+import { addSuccessNotification } from "Support/utils/appBuilder/querymanager/queries";
 
 const data = {};
 data.firstName = fake.firstName.toLowerCase().replaceAll("[^A-Za-z]", "");
