+++ cypress-tests/cypress/e2e/happyPath/workflows/eeTestcases/workflowWIthDatasource.cy.js
@@ -2,7 +2,7 @@ import { fake } from "Fixtures/fake";
 import { commonSelectors } from "Selectors/common";
 import { postgreSqlSelector } from "Selectors/postgreSql";
 import { postgreSqlText } from "Texts/postgreSql";
-import { deleteWorkflowAndDS, deleteDatasource } from "Support/utils/dataSource";
+import { deleteWorkflowAndDS, deleteDatasource } from "Support/utils/marketplace/datasources/dataSource";
 import { dataSourceSelector } from "Selectors/dataSource";
 import { harperDbText } from "Texts/harperDb";
 import { workflowsText } from "Texts/workflows";
@@ -10,7 +10,7 @@ import { workflowSelector } from "Selectors/workflows";
 import {
   fillDataSourceTextField,
   selectAndAddDataSource,
-} from "Support/utils/postgreSql";
+} from "Support/utils/marketplace/datasources/postgreSql";
 import {
   enterJsonInputInStartNode,
   verifyTextInResponseOutputLimited,
