+++ cypress-tests/cypress/support/utils/appBuilder/querymanager/queries.js
